requirements.txt
rich
pytest
torch>=2
torchaudio
numpy
setuptools
proto-plus
jinja2
loguru
python-dotenv
bittensor==7.4.0
nltk
scikit-learn
librosa
transformers
IPython
notebook
accelerate
bitsandbytes
pydub
sentencepiece
python-multipart
llama-cpp-python
unsloth
